chore(main): inline locate_file
It was only used in one place, no reason keeping a full module with
custom error messages around. We can use anyhow and option, and a
better function name while we’re at it.
